Updated php classes
CPE, facilimanagement: if not logged in, return to log in screen. If the
role isn't set to "CPE" then redirected to appropriate location
updated addemployee to include cpe functionality designed solely for the
CPE. in addition to creating a new employee, a cpe can create a new
manager and insert into a site the new given employee.
newManager, CPEManagement: empty file, might delete. dont care