update files
Please note when referring to employees, I am talking about both
employees and managers.
From here a CPE can decide to either modify existing employees or create
new employees. Once the form has been submitted, this will redirect the
cpe to cpeExistingEmployee,

cpeExistingEmployee.php:
This will allow the CPE to either terminate the employee or modify
limited employee information

terminate: doesn't delete from the employee database, but rather sets
the enddate to the current time and changes the role to 'NLE' as in no
longer employed. the employee ID is removed from the employeelist
database
modify: ability to relocate employee to another facility or change
roles. if the employee is relocated to another facility, then the entry
for that employee in the supervises table is dropped(as he no longer
works there).